* Added some helpful comments to Unit Morph hooks, Building Morph hooks, and Unit Tooltip hooks. * Removed some hook functions in hooks/unit_tooltip.cpp that was cluttering the code a bit too much.

+++ trunk/GPTP/src/hooks/unit_tooltip_inject.cpp
@@ -32,7 +32,7 @@ void __stdcall statdatalUserInteractWrapper(BinDlg *consoleIcon) {
   switch (gui->overlayType) {
     case 0:
       createContextHelpFromDialog(consoleIcon,
-        hooks::getWeaponTooltipString(gui->id, *activePortraitUnit));
+        hooks::getWeaponTooltipString((u8) gui->id, *activePortraitUnit));
       break;
     case 1:
       createContextHelpFromDialog(consoleIcon,
